[chore] Make ErrorReportSanitizer a func instead of interface
This way, there's no need for a type to own the function, and it can
more easily be replaced in tests etc. It also limits the package
namespace population in this package as there will be an analogous
interface for sessions, and instead of needing two interfaces (or
wondering if they should be two methods on the same interface).