[feat] Sanitizers return errors instead of context
This has the advantage that users don't get confused and try and return
a ctx that is derived from the ctx in the input parameter to the
sanitizers. This input context is often tied to a request in the user's
applications, and thus could lead to confusion about why errors aren't
being reported to Busgnag (the HTTP request's context was cancelled).

It *does* remove any control that the user previously had over the
deadline etc. of the Bugsnag payload. In the long term, there are other
ways of re-introducing this feature, should it be necessary, without
breaking backwards compatibility.

Additional changes:
- Simplifies internal structure by removing the `report` type.